Rename ParsingContext to ContractFlattener
