+++ packages/discovery/src/discovery/output/saveDiscoveryResult.ts
@@ -4,7 +4,7 @@ import { mkdirp } from 'mkdirp'
 import path, { dirname, posix } from 'path'
 import { rimraf } from 'rimraf'
 
-import { ContractFlattener } from '../../flatten/ParsingContext'
+import { ContractFlattener } from '../../flatten/ContractFlattener'
 import { EthereumAddress } from '../../utils/EthereumAddress'
 import { Analysis } from '../analysis/AddressAnalyzer'
 import { DiscoveryConfig } from '../config/DiscoveryConfig'
